requirements.txt
absl-py==1.3.0
aioice==0.7.6
aiortc==1.3.2
altair==4.2.0
apturl==0.5.2
asttokens==2.2.1
attrs==22.2.0
av==9.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base==1.0.4
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
grpcio==1.51.1
httplib2==0.14.0
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
monotonic==1.5
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycocotools==2.0.6
pycparser==2.21
pycups==1.9.73
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
pylibsrtp==0.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.9
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==13.0.1
rsa==4.9
scipy==1.10.0
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
semver==2.13.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==5.0.0
ssh-import-id==5.10
stack-data==0.6.2
system-service==0.3
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.17.1
wadllib==1.3.3
watchdog==2.2.1
wcwidth==0.2.5
Werkzeug==2.2.2
xkit==0.0.0
zipp==3.11.0
